A few more small tidy-ups
